Move tests of module preload.
